+++ AdventOfCode2020/Day06/Solution06.cs
@@ -11,7 +11,7 @@ namespace AdventOfCode2020.Day06
 
         #region Problem One
 
-        public static async Task<int> ProblemOneAsync(IEnumerable<IEnumerable<string>> blocks = null)
+        public static async Task<long> ProblemOneAsync(IEnumerable<IEnumerable<string>> blocks = null)
         {
             blocks ??= await ReadInputAsync();
             return blocks.Sum(CountUniqueCharactersInLines);
@@ -27,7 +27,7 @@ namespace AdventOfCode2020.Day06
 
         #region Problem Two
 
-        public static async Task<int> ProblemTwoAsync(IEnumerable<IEnumerable<string>> blocks = null)
+        public static async Task<long> ProblemTwoAsync(IEnumerable<IEnumerable<string>> blocks = null)
         {
             blocks ??= await ReadInputAsync();
             return blocks.Sum(CountSameAnsweredQuestions);
